Move reusable sender code out of json-based ocm share manager
